+++ tests/general/default_config_test.go
@@ -58,7 +58,7 @@ func TestDefaultGatewayConfig(t *testing.T) {
 			config := collector.EffectiveConfig(t)
 			require.Equal(t, map[string]any{
 				"exporters": map[string]any{
-					"otlphttp": map[string]any{
+					"otlp_http": map[string]any{
 						"traces_endpoint": "https://ingest.not.real.signalfx.com/v2/trace/otlp",
 						"sending_queue": map[string]any{
 							"num_consumers": 32,
@@ -94,7 +94,7 @@ func TestDefaultGatewayConfig(t *testing.T) {
 						"token":            "<redacted>",
 						"log_data_enabled": false,
 					},
-					"otlphttp/entities": map[string]any{
+					"otlp_http/entities": map[string]any{
 						"logs_endpoint": "https://ingest.not.real.signalfx.com/v3/event",
 						"headers": map[string]any{
 							"X-SF-Token": "<redacted>",
@@ -124,6 +124,14 @@ func TestDefaultGatewayConfig(t *testing.T) {
 						},
 						"ingress": map[string]any{
 							"endpoint": fmt.Sprintf("%s:6060", ip),
+						},
+					},
+					"http_forwarder/signalfx": map[string]any{
+						"egress": map[string]any{
+							"endpoint": "https://ingest.not.real.signalfx.com",
+						},
+						"ingress": map[string]any{
+							"endpoint": fmt.Sprintf("%s:9943", ip),
 						},
 					},
 					"zpages": map[string]any{
@@ -197,9 +205,6 @@ func TestDefaultGatewayConfig(t *testing.T) {
 							},
 						},
 					},
-					"signalfx": map[string]any{
-						"endpoint": fmt.Sprintf("%s:9943", ip),
-					},
 					"zipkin": map[string]any{
 						"endpoint": fmt.Sprintf("%s:9411", ip),
 					},
@@ -217,20 +222,15 @@ func TestDefaultGatewayConfig(t *testing.T) {
 					},
 				},
 				"service": map[string]any{
-					"extensions": []any{"headers_setter", "health_check", "http_forwarder", "zpages"},
+					"extensions": []any{"headers_setter", "health_check", "http_forwarder", "http_forwarder/signalfx", "zpages"},
 					"pipelines": map[string]any{
 						"logs": map[string]any{
 							"exporters":  []any{"splunk_hec", "splunk_hec/profiling"},
 							"processors": []any{"memory_limiter", "batch"},
 							"receivers":  []any{"routing/logs"},
 						},
-						"logs/signalfx": map[string]any{
-							"exporters":  []any{"signalfx"},
-							"processors": []any{"memory_limiter", "batch"},
-							"receivers":  []any{"signalfx"},
-						},
 						"logs/entities": map[string]any{
-							"exporters":  []any{"otlphttp/entities"},
+							"exporters":  []any{"otlp_http/entities"},
 							"processors": []any{"memory_limiter", "batch"},
 							"receivers":  []any{"routing/logs"},
 						},
@@ -241,7 +241,7 @@ func TestDefaultGatewayConfig(t *testing.T) {
 						"metrics": map[string]any{
 							"exporters":  []any{"signalfx"},
 							"processors": []any{"memory_limiter", "batch"},
-							"receivers":  []any{"otlp", "signalfx"},
+							"receivers":  []any{"otlp"},
 						},
 						"metrics/internal": map[string]any{
 							"exporters":  []any{"signalfx/internal"},
@@ -249,7 +249,7 @@ func TestDefaultGatewayConfig(t *testing.T) {
 							"receivers":  []any{"prometheus/internal"},
 						},
 						"traces": map[string]any{
-							"exporters":  []any{"otlphttp"},
+							"exporters":  []any{"otlp_http"},
 							"processors": []any{"memory_limiter", "batch"},
 							"receivers":  []any{"jaeger", "otlp", "zipkin"},
 						},
@@ -296,7 +296,7 @@ func TestDefaultAgentConfig(t *testing.T) {
 					"debug": map[string]any{
 						"verbosity": "detailed",
 					},
-					"otlp/gateway": map[string]any{
+					"otlp_grpc/gateway": map[string]any{
 						"endpoint": ":4317",
 						"tls": map[string]any{
 							"insecure": true,
@@ -305,7 +305,7 @@ func TestDefaultAgentConfig(t *testing.T) {
 							"authenticator": "<redacted>",
 						},
 					},
-					"otlphttp": map[string]any{
+					"otlp_http": map[string]any{
 						"headers": map[string]any{
 							"X-SF-Token": "<redacted>",
 						},
@@ -333,7 +333,7 @@ func TestDefaultAgentConfig(t *testing.T) {
 						"token":            "<redacted>",
 						"log_data_enabled": false,
 					},
-					"otlphttp/entities": map[string]any{
+					"otlp_http/entities": map[string]any{
 						"logs_endpoint": "https://ingest.not.real.signalfx.com/v3/event",
 						"headers": map[string]any{
 							"X-SF-Token": "<redacted>",
@@ -481,14 +481,14 @@ func TestDefaultAgentConfig(t *testing.T) {
 							"receivers":  []any{"prometheus/internal"},
 						},
 						"traces": map[string]any{
-							"exporters":  []any{"otlphttp", "signalfx"},
+							"exporters":  []any{"otlp_http", "signalfx"},
 							"processors": []any{"memory_limiter", "batch", "resourcedetection"},
 							"receivers":  []any{"jaeger", "otlp", "zipkin"},
 						},
 						"logs/entities": map[string]any{
 							"receivers":  []any{"nop"},
 							"processors": []any{"memory_limiter", "batch", "resourcedetection"},
-							"exporters":  []any{"otlphttp/entities"},
+							"exporters":  []any{"otlp_http/entities"},
 						},
 					},
 				},