+++ tests/receivers/scriptedinputs/script_test.go
@@ -129,6 +129,7 @@ func TestScriptReceiverLsof(t *testing.T) {
 }
 
 func TestScriptReceiverNetstat(t *testing.T) {
+	t.Skip("skip flaky test on deprecated code")
 	tc := testutils.NewTestcase(t)
 	defer tc.PrintLogsOnFailure()
 	defer tc.ShutdownOTLPReceiverSink()
@@ -145,26 +146,6 @@ func TestScriptReceiverNetstat(t *testing.T) {
 
 		lr := receivedOTLPLogs[0].ResourceLogs().At(0).ScopeLogs().At(0).LogRecords().At(0)
 		assert.Regexp(tt, regexp.MustCompile("Proto\\s+Recv-Q\\s+Send-Q\\s+LocalAddress\\s+ForeignAddress\\s+State"), lr.Body().Str())
-	}, 10*time.Second, 10*time.Millisecond, "Failed to receive expected logs")
-}
-
-func TestScriptReceiverOpenPorts(t *testing.T) {
-	tc := testutils.NewTestcase(t)
-	defer tc.PrintLogsOnFailure()
-	defer tc.ShutdownOTLPReceiverSink()
-
-	_, shutdown := tc.SplunkOtelCollectorProcess("script_config_openPorts.yaml")
-	defer shutdown()
-
-	assert.EventuallyWithT(t, func(tt *assert.CollectT) {
-		if tc.OTLPReceiverSink.LogRecordCount() == 0 {
-			assert.Fail(tt, "no logs received")
-			return
-		}
-		receivedOTLPLogs := tc.OTLPReceiverSink.AllLogs()
-
-		lr := receivedOTLPLogs[0].ResourceLogs().At(0).ScopeLogs().At(0).LogRecords().At(0)
-		assert.Regexp(tt, regexp.MustCompile("Proto\\s+Port"), lr.Body().Str())
 	}, 10*time.Second, 10*time.Millisecond, "Failed to receive expected logs")
 }
 
@@ -185,26 +166,6 @@ func TestScriptReceiverPackage(t *testing.T) {
 
 		lr := receivedOTLPLogs[0].ResourceLogs().At(0).ScopeLogs().At(0).LogRecords().At(0)
 		assert.Regexp(tt, regexp.MustCompile("NAME\\s+VERSION\\s+RELEASE\\s+ARCH\\s+VENDOR\\s+GROUP"), lr.Body().Str())
-	}, 10*time.Second, 10*time.Millisecond, "Failed to receive expected logs")
-}
-
-func TestScriptReceiverProtocol(t *testing.T) {
-	tc := testutils.NewTestcase(t)
-	defer tc.PrintLogsOnFailure()
-	defer tc.ShutdownOTLPReceiverSink()
-
-	_, shutdown := tc.SplunkOtelCollectorProcess("script_config_protocol.yaml")
-	defer shutdown()
-
-	assert.EventuallyWithT(t, func(tt *assert.CollectT) {
-		if tc.OTLPReceiverSink.LogRecordCount() == 0 {
-			assert.Fail(tt, "no logs received")
-			return
-		}
-		receivedOTLPLogs := tc.OTLPReceiverSink.AllLogs()
-
-		lr := receivedOTLPLogs[0].ResourceLogs().At(0).ScopeLogs().At(0).LogRecords().At(0)
-		assert.Regexp(tt, regexp.MustCompile("IPdropped\\s+TCPrexmits\\s+TCPreorder\\s+TCPpktRecv\\s+TCPpktSent\\s+UDPpktLost\\s+UDPunkPort\\s+UDPpktRecv\\s+UDPpktSent"), lr.Body().Str())
 	}, 10*time.Second, 10*time.Millisecond, "Failed to receive expected logs")
 }